quizlive-backend/quizlive_backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
psycopg2-binary==2.9.9
dj-database-url==2.2.0
redis==5.0.8
django-redis==5.4.0
hiredis==3.0.0
django-environ==0.11.2
whitenoise==6.7.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0
drf-spectacular==0.27.2
pytest==8.3.2
pytest-django==4.8.0
pytest-asyncio==0.23.8
factory-boy==3.3.1
ruff==0.5.7
